Refactors asset generation and removes unused components
Simplifies asset generation logic by introducing new abstractions for
layout, selection, and placement. Refines asset dependency tracking and
streamlines property handling across asset schemas.

Enhances entity spawning with improved player handling and spawn point
management. Updates event constants for clarity and consistency.
Eliminates outdated selection types and replaces them with more robust
selection utilities.

Addresses code cleanliness and maintainability by removing redundant
methods and fields, consolidating asset initialization logic, and
replacing direct identifier references with entry-based handling.

+++ src/main/java/com/transcendruins/assets/rendermaterials/RenderMaterialInstance.java
@@ -56,7 +56,7 @@ public final class RenderMaterialInstance extends AssetInstance {
 
     private int createMaterialBitMask() {
 
-        byte mask = 0;
+        int mask = 0;
         boolean[] bits = { lit, diffuses, hasAmbient, hasSpecular, hasSpecularMap, transparent, hasBackfaceCulling };
 
         for (int i = 0; i < bits.length; i++) {
